Refactor `InternetGateway` tests to use fixture
Refactoring e2e tests to use pytest.fixture for the resource under
test. Prior to these changes, assertion failures could leave dangling
resources in the testing account.
